ModelTesting

Documentation for ModelTesting.

ModelTesting is intended to facilitate the easy testing and post-facto analysis of ModelingToolkit models. It currently provides two key bits of functionality:

Serialization of MTK solutions into a common format through discretize_solution.

Comparison of solutions to each other and to previously-saved data through compare.

ModelTesting.DefaultComparison

ModelTesting.DefaultComparison

ModelTesting.compare

ModelTesting.compare

ModelTesting.discretize_solution

ModelTesting.test_instantaneous

ModelTesting.DefaultComparison — Type

DefaultComparison is a convenience default comparison method for solutions and reference data.

By default it's instantiated with l∞, l1, l2, RMS, and final l1 comparsons. More can be added by passing a dict of name=>comparison function pairs into field_cmp, with the signature (delta, dtdelta, t) -> [vector difference]. The arguments of these comparison functions are delta for the raw difference beteween the value, `dtdelta` for the product of dt and delta (if you want to approximate the integral of the error), and the timestamps at which the data was produced. The arguments consist of a vector per compared field.

ModelTesting.DefaultComparison — Method

(d::DefaultComparison)(c, names, b, t, n, r)

Default comparison implementation. Returns a dataframe that describes the result of running each comparison operation on the given data.

#Arguments

c: The symbolic container from which the values we're comparing came

names: The names to output for the fields we're comparing (not always the same as the names of the fields being compared)

b: The symbolic variables that are being compared

t: The timestamps at which the comparison is taking place

n: The new values being compared

r: The reference values to compare against

ModelTesting.compare — Function

compare(
    new_sol::SciMLBase.AbstractTimeseriesSolution,
    reference::DataFrame,
    cmp=DefaultComparison(); to_name=string, warn_observed=true)

compare compares the results of new_sol to a result in reference, which may come from either experiment or a previous model execution. The format of reference is that produced by discretize_solution:

A column named "timestamp" that indicates the time the measurement was taken (referenced to the same timebase as the solution provided)

Columns with names matching the names in the completed system new_sol (fully qualified)

If new_sol is dense then the discretization nodes don't need to line up with the reference and the interpolant will be used instead. If it is not dense then the user must ensure that the saved states in are at the same times as the "timestamp"s are in the reference data.

compare will use the comparison method specified by cmp to compare the two solutions; by default it uses the DefaultComparison (which offers l1, l2, and rms comparisons per observed value), but you can pass your own. Look at DefaultComparison for more details on how.

The two optional named parameters are to_name (which is used to convert the symbolic variables in new_sol into valid column names for the dataframe) and warn_observed which controls whether compare complains about comparing observed values to non-observed values. We suggest leaving warn_observed on for regression testing (it'll give you a note when MTK changes the simplifcation of the system).

ModelTesting.compare — Function

compare(
    new_sol::SciMLBase.AbstractTimeseriesSolution,
    reference::DataFrame,
    over::Vector{<:Union{Pair{<:Any, String}, Pair{<:Any, Pair{String, String}}}}
    cmp=DefaultComparison(); warn_observed=true)

Lower level version of compare that lets you specify the mapping from values in new_sol (specified as either variable => name or variable => (inputname, outputname) pairs) to the column names in reference. The tuple-result version lets you specify what the column names are when passed to the comparison method. Since the mapping is explicit this version does not take to_name, but it still does take warn_observed. See the implicit-comparison version of compare for more information.

ModelTesting.discretize_solution — Method

discretize_solution(solution::SciMLBase.AbstractTimeseriesSolution[, time_ref::SciMLBase.AbstractTimeseriesSolution]; measured=nothing, all_observed=false)

discretize_solution takes a solution and an optional time reference and converts it into a dataframe. This dataframe will contain either:

The variables (unknowns or observeds) named in measured, if provided.

The variables marked as measured if all_observed is false and measured is nothing.

All variables in the system if all_observed is true and measured is nothing.

The dataframe will contain a column called timestamp for each of the discretization times and a column for each observed value.

If no time reference is provided then the timebase used to discretize solution will be used instead.

ModelTesting.test_instantaneous — Method

test_instantaneous(sys::ModelingToolkit.AbstractSystem,
    ic,
    checks::Union{Num, Array};
    t = nothing)

test_instantaneous is a helper wrapper around constructing and executing an ODEProblem at a specific condition. It's intended for use in basic sanity checking of MTK models, for example to ensure that the derivative at a given condition has the correct sign. It should be passed the system, the initial condition dictionary to be given to the ODEProblem constructor (including parameters), and a list of observable values from the system to extract from the ODEProblem. If checks is a single Num then the result will be that observed value; otherwise it will return an array in the same order as the provided checks.